# infrastructure/rest-api/trading-control-service-open-api/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Trading Control Service API
  version: 1.0.0

paths:
  /api/v1/markets:
    get:
      summary: Get supported markets and stream channel capabilities
      operationId: getMarkets
      tags:
        - markets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./schemas/market/Exchange.yaml"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/markets/{exchange}/{marketType}/instruments:
    parameters:
      - $ref: "#/components/parameters/Exchange"
      - $ref: "#/components/parameters/MarketType"
    get:
      summary: Get instruments for a given exchange and market type
      operationId: getMarketInstruments
      tags:
        - markets
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "./schemas/instrument/MarketInstrumentsResponse.yaml"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/streams:
    get:
      summary: Get configured streams
      operationId: getStreams
      tags:
        - streams
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "./schemas/stream/StreamResponse.yaml"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      summary: Create a stream configuration
      operationId: createStream
      tags:
        - streams
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/stream/CreateStreamRequest.yaml"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "./schemas/stream/StreamResponse.yaml"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  /api/v1/streams/{streamId}:
    parameters:
      - $ref: "#/components/parameters/StreamId"
    get:
      summary: Get a stream by id
      operationId: getStream
      tags:
        - streams
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "./schemas/stream/StreamResponse.yaml"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      summary: Update a stream's desired spec
      description: Set desiredState to ENABLED/DISABLED to start/stop the stream, and/or replace the channel set.
      operationId: updateStream
      tags:
        - streams
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/stream/UpdateStreamRequest.yaml"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "./schemas/stream/StreamResponse.yaml"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      summary: Delete a configured stream
      operationId: deleteStream
      tags:
        - streams
      responses:
        "204":
          description: No Content
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  parameters:
    Exchange:
      name: exchange
      in: path
      required: true
      schema:
        type: string
    MarketType:
      name: marketType
      in: path
      required: true
      schema:
        type: string
    StreamId:
      name: streamId
      in: path
      required: true
      schema:
        type: string

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "./schemas/error/ErrorResponse.yaml"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "./schemas/error/ErrorResponse.yaml"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "./schemas/error/ErrorResponse.yaml"